+++ RELEASENOTES.md
@@ -77,6 +77,7 @@
     *   Increase required min API version to 21.
     *   `TransformationException` is now used to describe errors that occur
         during a transformation.
+    *   Add `TransformationRequest` for specifying the transformation options.
 *   MediaSession extension:
     *   Remove deprecated call to `onStop(/* reset= */ true)` and provide an
         opt-out flag for apps that don't want to clear the playlist on stop.